Play a five round game and keep score
I had a problem in the previous commit getting a different throw from
the computer each time I played a new round. The solution turned out to
be that every time I call the function that makes the computer throw, I
have to reassign the variable to it that passes the throw to the round.

In other words: Every time I call getComputerChoice(), I have to
reassign computerSelection to it.